Разбил код в main на утилитарные ф-ции (в util) и код, генериующий данные (data)
